+++ BlueprintUILists/Sources/List.swift
@@ -140,12 +140,12 @@ extension List {
                     constraint.maximum
                 }
                 
-            case .measureContent(let horizontalFill, let verticalFill, let limit):
-                return ElementContent() { constraint -> CGSize in
+            case .measureContent(let horizontalFill, let verticalFill, let safeArea, let limit):
+                return ElementContent() { constraint, environment -> CGSize in
                     let measurements = ListView.contentSize(
                         in: constraint.maximum,
                         for: self.properties,
-                        safeAreaInsets: .zero,
+                        safeAreaInsets: safeArea.safeArea(with: environment),
                         itemLimit: limit
                     )